refactor(settings): extract LanguagePickerOverlay into its own widget file
Move the self-contained `_LanguagePickerOverlay` StatefulWidget out of
general_settings_content.dart into new
new_settings_options/widgets/language_picker_overlay.dart, promoted to a
public `LanguagePickerOverlay`. Body is byte-identical (verified via
token-stream normalization vs main); the only deltas are the
private->public rename and the added `super.key` parameter.

The widget is autonomous — it takes `anchorOffset` + `currentLang` and
returns its result via `Navigator.pop`, reading no host state. The
`configProvider` read + result handling stay in the host's
`_showLanguagePicker`, which constructs the widget at the identical
`showGeneralDialog` call site (context ancestry unchanged).

Dropped 3 now-overlay-only host imports (sfx_service, game_service,
gamepad_nav). Host 955 -> 725; new file 234 lines. No behaviour change,
no public-API change to the host.
